+++ hrefs/__init__.py
@@ -1,6 +1,6 @@
 """Hyperlinks for pydantic models"""
 
-__version__ = "0.4"
+__version__ = "0.5"
 
 from .href import Href, Referrable
 from .model import BaseReferrableModel, PrimaryKey